requirements.txt
streamlit
pyyaml
pandas
sqlalchemy
psycopg[binary]
cryptography
